refactor: rename setInteger -> setLong
Consistent with how setDouble() accepts a Double even though
the Weaviate's data type is number.
PropertyBuilder deals with Java data types so the methods
should be named accordingly.

+++ src/main/java/io/weaviate/client6/v1/internal/orm/MapBuilder.java
@@ -25,7 +25,7 @@ public class MapBuilder implements PropertiesBuilder<Map<String, Object>> {
   }
 
   @Override
-  public void setInteger(String property, Long value) {
+  public void setLong(String property, Long value) {
     properties.put(property, value);
   }
